fix: build: testing mode should append paths properly, even with lockfiles
As `config.nims` is not loaded up by the compiler when tests are
executed (generally), lockfiles would skip append-path generation in
test builds while also ensuring the paths would not be appended
`neo.paths`.

This fixes it by generating append-paths in the lockfile codepath correctly when tests are being
compiled.